fixes #37
Previously would forward to /dashboard and then to if user wasn't sign in /login

Now instead it forwards to /login until user is authenticated, at which point they are forwarded to the dashboard. This feels more natural.

Also adds loading indicator for when auth hasn't finished figuring out who the user is yet